requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
tokenizers>=0.13.0
accelerate>=0.26.0

# Scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
huggingface-hub>=0.15.0

# Optional: for development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
